Revert "US121812 WIP avail dates component, fixing lesson header size"
This reverts commit ed0cbc02c7d97a3832e3fba2f97289cd44315671.
