+++ d2l-sequence-viewer/d2l-sequence-viewer-sidebar.js
@@ -33,7 +33,7 @@ class D2LSequenceViewerSidebar extends mixinBehaviors([
 				justify-content: center;
 				align-items: center;
 				background-color: var(--d2l-asv-primary-color);
-				min-width: 238px;
+				min-width: 250px;
 				padding: 8px 20px;
 			}
 			.m-module-heading:hover {
